+++ changelog.md
@@ -7,6 +7,13 @@
   - 🎵 Added a feature to show a second, blurred duplicate of the thumbnail image in the thumbnail overlay as a background.
   - Added a configuration option for the global alerts system.
   - Added Turkish translations (by [@kcangny](https://github.com/kcangny)).
+- **Plugin Changes:**
+  - Added new features to the BYTM object. [Refer to the API docs for details.](https://github.com/Sv443/BetterYTM/blob/develop/contributing.md#global-functions-and-classes)
+    - `loggers`: An object of predefined Logger instances of different categories.
+    - `Logger`: The Logger class, which can be used to create a new logging category. Any created Logger instance will automatically share its logs with BYTM's internal logging system.
+- **Internal Changes:**
+  - Added `Logger` class to tag every log with a category, in preparation for a future log filtering feature.
+  - Refactored logging system to use new `Logger` class instances.
 
 <div class="pr-link-cont">
   <a href="https://github.com/Sv443/BetterYTM/pull/TODO" rel="noopener noreferrer">Also see pull request #TODO</a>